+++ src/app/app.component.html
@@ -8,7 +8,7 @@
             <mat-icon>description</mat-icon>
             Order
         </a>
-        <a mat-button routerLink="/order-new" routerLinkActive="active-link">
+        <a mat-button routerLink="/order-form" routerLinkActive="active-link">
             <mat-icon>add_shopping_cart</mat-icon>
             New Order
         </a>
