+++ src/app/components/order-confirmation/order-confirmation.component.html
@@ -13,21 +13,33 @@
         <th>Pack Size</th>
         <th>Price</th>
         <th>Quantity</th>
-        <th>Delivery Date</th>
       </tr>
     </thead>
     <tbody>
       <tr *ngFor="let product of orderData.profiles">
         <td>{{ product.profileDescription }}</td>
-        <td>{{ product.unitTypePd }}</td>
-        <td>{{ product.packSizePd }}</td>
+        <td>{{ product.unitType }}</td>
+        <td>{{ product.packSize }}</td>
         <td class="price">{{ product.salesPrice | number:'1.2-2' }}</td>
         <td>{{ product.quantity }}</td>
-        <td>{{ orderData.deliveryDate | date: 'MM/dd/yyyy'}}</td>
       </tr>
     </tbody>
   </table>
-  <p><strong>Total Price:</strong> {{ orderData.totalPrice | currency:'USD':true }}</p>
+
+  <div>
+    <div class="total"><strong>Total:</strong>
+      <span id="total-amount">{{ orderData.totalPrice| currency:'USD':true }}</span>
+    </div>
+  </div>
+
+  <div>
+    <p>Delivery date: {{orderData.deliveryDate | date: 'MM/dd/yyyy'}}</p>
+  </div>
+  <div *ngIf="orderData.shipToId">
+    <p>Delivery location: {{orderData.shipToName}}</p>
+  </div>
+
+
   <p>Thank you,<br>Your Sales Team</p>
   <button class="button" (click)="goBack()">Back</button>
 </div>
