+++ src/app/components/order-exists/order-exists.component.css
@@ -24,12 +24,14 @@ body {
     align-items: center;
     padding: 10px;
     border-radius: 8px;
-    background-color: rgba(0, 16, 46, 1); /* Dark blue background */
+    background-color: rgba(0, 16, 46, 1);
+    /* Dark blue background */
 }
 
 .image-container img {
     max-width: 100%;
-    height: 100px; /* Adjust height to match the order page */
+    height: 100px;
+    /* Adjust height to match the order page */
     border-radius: 8px;
 }
 
@@ -58,7 +60,6 @@ ul li {
 .total {
     font-weight: bold;
     text-align: right;
-    margin-bottom: 20px;
     font-size: 18px;
 }
 
@@ -70,7 +71,8 @@ ul li {
     width: 100%;
     border-collapse: collapse;
     margin-bottom: 20px;
-    table-layout: fixed; /* Fixed table layout */
+    table-layout: fixed;
+    /* Fixed table layout */
 }
 
 .order-grid th,
@@ -78,39 +80,28 @@ ul li {
     border: 1px solid #ccc;
     padding: 8px;
     text-align: center;
-    word-wrap: break-word; /* Ensure content wraps */
+    word-wrap: break-word;
+    /* Ensure content wraps */
 }
 
 .order-grid th {
-    background-color: #d3d3d3; /* Light grey background for table headers */
-    color: #333; /* Dark grey text color for table headers */
+    background-color: #d3d3d3;
+    /* Light grey background for table headers */
+    color: #333;
+    /* Dark grey text color for table headers */
     font-weight: bold;
-    font-size: 14px; /* Match the inventory grid font size */
+    font-size: 14px;
+    /* Match the inventory grid font size */
 }
 
 .order-grid td {
-    font-size: 14px; /* Match the inventory grid font size */
+    font-size: 14px;
+    /* Match the inventory grid font size */
 }
 
-.hidden-column {
-    display: none;
-}
 
-/* Set specific column widths */
-.order-grid th:nth-child(1),
-.order-grid td:nth-child(1),
-.order-grid th:nth-child(2),
-.order-grid td:nth-child(2) {
-    display: none; /* Hide Profile ID and Customer ID columns */
-}
 
-.order-grid th:nth-child(3) {
-    width: 30%; /* Other columns */
-}
 
-.order-grid td input[type="number"] {
-    width: 70px;
-}
 
 @media (max-width: 600px) {
     .container {
@@ -122,7 +113,8 @@ ul li {
         font-size: 24px;
     }
 
-    p, ul {
+    p,
+    ul {
         font-size: 16px;
     }
 
@@ -137,17 +129,25 @@ ul li {
     }
 
     .order-grid {
-        width: 100%; /* Full width for grid on mobile */
-        margin: 0; /* Remove margin */
-        border: 0; /* Remove border */
-        display: table; /* Ensure proper table layout */
+        width: 100%;
+        /* Full width for grid on mobile */
+        margin: 0;
+        /* Remove margin */
+        border: 0;
+        /* Remove border */
+        display: table;
+        /* Ensure proper table layout */
     }
 
     .container {
-        width: 100%; /* Full width for container on mobile */
-        margin: 0; /* Remove margin */
-        padding: 0.5; /* Remove padding */
-        border: 0; /* Remove border */
+        width: 100%;
+        /* Full width for container on mobile */
+        margin: 0;
+        /* Remove margin */
+        padding: 0.5;
+        /* Remove padding */
+        border: 0;
+        /* Remove border */
     }
 }
 
@@ -176,5 +176,8 @@ ul li {
     padding: 5px 10px;
     cursor: pointer;
     border-radius: 5px;
-    }
-    
+}
+
+.no-wrap {
+    white-space: nowrap;
+}
