+++ src/app/components/order-exists/order-exists.component.html
@@ -9,35 +9,41 @@
   <table class="order-grid" id="order-grid">
     <thead>
       <tr>
-        <th>Customer Name</th>
-        <th>Sales Rep</th>
         <th>Product Description</th>
         <th>Unit Type</th>
         <th>Pack Size</th>
         <th>Price</th>
         <th>Quantity</th>
-        <th>Delivery Date</th>
-        <th *ngIf="orderData.id">Shipto Name</th>
       </tr>
     </thead>
     <tbody>
       <tr *ngFor="let profile of orderData.profiles">
-        <td>{{ orderData.customerName }}</td>
-        <td>{{ orderData.salesRepName }}</td>
         <td>{{ profile.profileDescription }}</td>
-        <td>{{ profile.unitTypePd }}</td>
-        <td>{{ profile.packSizePd }}</td>
+        <td>{{ profile.unitType }}</td>
+        <td>{{ profile.packSize }}</td>
         <td class="price">{{ profile.salesPrice| currency:'USD':true }}</td>
         <td>{{ profile.quantity }}</td>
-        <td>{{ orderData.deliveryDate | date: 'MM/dd/yyyy' }}</td>
-        <td *ngIf="orderData.id">{{ orderData.shipToName }}</td>
       </tr>
     </tbody>
   </table>
-  <div style="margin-top: 20px;">
+
+  <div>
     <div class="total"><strong>Total:</strong>
       <span id="total-amount">{{ orderData.totalPrice| currency:'USD':true }}</span>
     </div>
   </div>
+
+  <div>
+    <p>Sales rep: {{orderData.salesRepName}}</p>
+  </div>
+  <div>
+    <p>Delivery date: {{orderData.deliveryDate | date: 'MM/dd/yyyy'}}</p>
+  </div>
+  <div *ngIf="orderData.shipToId">
+    <p>Delivery location: {{orderData.shipToName}}</p>
+  </div>
+
+
+
   <button class="button" (click)="goBack()">Back</button>
 </div>
