+++ src/app/components/order-links/order-links.component.html
@@ -36,7 +36,7 @@
       <tbody>
         <tr *ngFor="let customer of filteredCustomers;">
           <td>{{ customer.name }}</td>
-          <td><a [routerLink]="['/customer', customer.id, 'order-new']">Order Form</a></td>
+          <td><a [routerLink]="['/customer', customer.id, 'order-form']">Order Form</a></td>
           <td><a href="javascript:void(0)" (click)="copyLink(customer.id)" (keyup)="copyLink(customer.id)">Copy Link</a></td>
         </tr>
       </tbody>
