--- src/app/components/order-new/order-new.component.css
@@ -1,217 +1,0 @@
-body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-
-
-.container {
-  width: 90%;
-  max-width: 800px;
-  margin: 20px auto;
-  padding: 20px;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-.page-title {
-  font-size: 28px;
-  text-align: center;
-  margin: 10px 0;
-  text-decoration: none;
-}
-
-.customer-sales-info {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 18px; /* Increased font size */
-  margin-bottom: 10px;
-}
-
-.sales-rep-info {
-  margin-left: auto;
-}
-
-.shipto-dropdown {
-  margin-top: 5px; /* Moved higher */
-  display: flex;
-  align-items: center;
-  font-size: medium;
-}
-
-.shipto-dropdown label {
-  margin-right: 10px; /* Added white space */
-}
-
-.input-select {
-  width: auto; /* Only as wide as needed */
-  padding: 4px; /* Less tall */
-}
-
-.section-title {
-  font-size: 20px; /* Increased font size */
-  margin-top: 20px;
-  margin-bottom: 10px;
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-bottom: 20px;
-  table-layout: fixed;
-}
-
-.table th,
-.table td {
-  border: 1px solid #ccc;
-  padding: 8px;
-  text-align: left;
-  word-wrap: break-word;
-}
-
-.table th {
-  background-color: #d3d3d3;
-  color: #333;
-  font-weight: bold;
-  font-size: 14px;
-}
-
-.table td {
-  font-size: 14px;
-}
-
-.table th:nth-child(1),
-.table td:nth-child(1) {
-  width: 40%; /* Wider description column */
-}
-
-.hidden-column {
-  display: none;
-}
-
-.error {
-  color: red;
-}
-
-.submit-btn {
-  width: 100%;
-  font-size: large;
-}
-
-.input-text,
-.input-select,
-.input-date,
-.input-number {
-  width: 100%;
-  padding: 8px;
-  margin: 4px 0;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.input-select {
-  width: auto;
-  padding: 4px;
-}
-
-.total-price {
-  margin-top: 0px;
-}
-
-.delivery-date {
-  margin-top: 5px;
-}
-
-.image-container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 10px;
-  border-radius: 8px;
-  background-color: rgba(0, 16, 46, 1);
-}
-
-.company-logo {
-  max-width: 100%;
-  height: 100px;
-  border-radius: 8px;
-}
-
-.input-number {
-  width: 100%;
-  padding: 2px; /* Adjust the padding to make it less tall */
-  margin: 3px 0;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-
-
-
-@media (max-width: 600px) {
-  .table th,
-  .table td {
-    font-size: 14px;
-    word-wrap: break-word;
-  }
-
-  .delivery-date {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .delivery-date label {
-    margin-bottom: 5px;
-    width: 100%;
-  }
-
-  .delivery-date input[type="date"],
-  .delivery-date input[type="text"] {
-    width: 100%;
-  }
-
-  .container {
-    width: 95%;
-    margin: 5px;
-  }
-
-  .input-text,
-  .input-select,
-  .input-date,
-  .input-number {
-    width: 100%;
-  }
-
-  .table {
-    width: 100%;
-    margin: 0;
-    border: 0;
-    display: table;
-  }
-
-  .container {
-    width: 100%;
-    margin: 0;
-    padding: 0.5;
-    border: 0;
-  }
-}
-
-
-.table th,
-.table td {
-  text-align: center; /* Center text in table headers and data cells */
-}
-
-.total-price {
-  font-weight: bold; /* Make text bold */
-  text-align: right; /* Align text to the right */
-  font-size: medium;
-}
