--- src/app/components/order-new/order-new.component.html
@@ -1,122 +1,0 @@
-<form [formGroup]="orderForm" (ngSubmit)="onSubmit()" class="order-form">
-    <div class="container">
-        <app-logo [companyId]="order.companyId"></app-logo>
-        <h2 class="page-title">Order Form</h2>
-
-        <div *ngIf="order.customerName && order.salesRepName" class="customer-sales-info">
-            <p><strong>Customer:</strong> {{ order.customerName }}</p>
-            <div class="sales-rep-info">
-                <strong>Sales Rep:</strong> {{ order.salesRepName }} <a [href]="'tel:' + order.salesRepPhone">{{ order.salesRepPhone }}</a>
-            </div>
-        </div>
-
-        @if (order.shipTos) {
-        <div class="shipto-dropdown">
-            <label for="ship-to"><strong>Ship to:</strong></label>
-            <select id="ship-to" formControlName="shipToId" class="input-select">
-                <option *ngFor="let shipTo of order.shipTos" [value]="shipTo.id">{{ shipTo.shipToName }}</option>
-            </select>
-            <div *ngIf="(orderForm.get('shipToId')?.touched || orderForm.get('shipToId')?.dirty) && orderForm.get('shipToId')?.errors?.['required']" class="error">
-                Ship to is required.
-            </div>
-        </div>
-        }
-
-        <h3 class="section-title">Specials</h3>
-        <div formArrayName="profiles" class="profiles">
-            <table class="table">
-                <thead>
-                    <tr>
-                        <th>Description</th>
-                        <th>Quantity</th>
-                        <th>Unit Type</th>
-                        <th>Pack Size</th>
-                        <th>Price</th>
-                        <th>Total</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <ng-container *ngFor="let profile of profileControls.controls; let i = index">
-                        <tr *ngIf="order.profiles[i].isSpecial" [formGroupName]="i" [ngStyle]="{'background-color': isQuantityEntered(i) ? 'lightgrey' : 'white'}">
-                            <input type="text" formControlName="profileDid" class="input-text hidden-column">
-                            <td>{{ order.profiles[i].profileDescription }}</td>
-                            <td><input type="number" formControlName="quantity" class="input-number">
-                                <div *ngIf="profile.get('quantity')?.errors?.['min']" class="error">The number must be greater than zero.</div>
-                            </td>
-                            <td>{{ order.profiles[i].unitTypePd }}</td>
-                            <td>{{ order.profiles[i].packSizePd }}</td>
-                            <td>{{ order.profiles[i].salesPrice | currency:'USD':'symbol':'1.2-2'}}</td>
-                            <td>{{ getRowTotalPrice(i) | currency:'USD':'symbol':'1.2-2' }}</td>
-                        </tr>
-                    </ng-container>
-                </tbody>
-            </table>
-        </div>
-
-        <h3 class="section-title">Products</h3>
-        <div formArrayName="profiles" class="profiles">
-            <table class="table">
-                <thead>
-                    <tr>
-                        <th>Description</th>
-                        <th>Quantity</th>
-                        <th>Unit Type</th>
-                        <th>Pack Size</th>
-                        <th>Price</th>
-                        <th>Total</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <ng-container *ngFor="let profile of profileControls.controls; let i = index">
-                        <tr *ngIf="!order.profiles[i].isSpecial" [formGroupName]="i" [ngStyle]="{'background-color': isQuantityEntered(i) ? 'lightgrey' : 'white'}">
-                            <input type="text" formControlName="profileDid" class="input-text hidden-column">
-                            <td>{{ order.profiles[i].profileDescription }}</td>
-                            <td><input type="number" formControlName="quantity" class="input-number">
-                                <div *ngIf="profile.get('quantity')?.errors?.['min']" class="error">The number must be greater than zero.</div>
-                            </td>
-                            <td>{{ order.profiles[i].unitTypePd }}</td>
-                            <td>{{ order.profiles[i].packSizePd }}</td>
-                            <td>{{ order.profiles[i].salesPrice | currency:'USD':'symbol':'1.2-2'}}</td>
-                            <td>{{ getRowTotalPrice(i) | currency:'USD':'symbol':'1.2-2' }}</td>
-                        </tr>
-                    </ng-container>
-                </tbody>
-            </table>
-            <div *ngIf="submitted && orderForm.get('profiles')?.hasError('atLeastOneQuantity')" class="error">
-                At least one quantity is required
-            </div>
-        </div>
-
-        <div *ngIf="profileControls.touched && profileControls.errors?.['atLeastOneQuantity']" class="error">
-            Enter at least one quantity
-        </div>
-
-        <div class="total-price">
-            <br>Total price: {{totalPrice | currency:'USD':'symbol':'1.2-2'}}
-            <div *ngIf="totalPrice > 10000" class="error">The total has to be less than $10,000</div>
-        </div>
-
-        <div class="delivery-date">
-            <label for="delivery-date"><strong>Delivery Date:</strong></label>
-            <input id="delivery-date" formControlName="deliveryDate" type="date" class="input-date">
-            <div *ngIf="formControls['deliveryDate'].touched && formControls['deliveryDate'].errors?.['required']" class="error">
-                Delivery date is required.
-            </div>
-            <div *ngIf="formControls['deliveryDate'].dirty && formControls['deliveryDate'].errors?.['dateAfterTomorrow']" class="error">
-                Delivery date must be after today.
-            </div>
-            <div *ngIf="formControls['deliveryDate'].dirty && formControls['deliveryDate'].errors?.['dateWithinThreeMonths']" class="error">
-                Delivery date must be within 3 months.
-            </div>
-            <div *ngIf="formControls['deliveryDate'].dirty && formControls['deliveryDate'].errors?.['dateNotOnSunday']" class="error">
-                Delivery date cannot be on Sunday.
-            </div>
-        </div>
-
-        <br><button type="submit" [disabled]="!orderForm.valid" class="submit-btn">Submit</button>
-    </div>
-
-    <br>Data to send to spring boot:
-    <br>
-    <pre>{{dataToBeSubmitted | json}}</pre>
-</form>
