--- src/app/components/order-new/order-new.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { OrderNewComponent } from './order-new.component';
-
-describe('OrderNewComponent', () => {
-  let component: OrderNewComponent;
-  let fixture: ComponentFixture<OrderNewComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [OrderNewComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(OrderNewComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
