+++ src/app/models/profile-confirmation.ts
@@ -1,0 +1,8 @@
+export interface ProfileConfirmation {
+  profileDescription: string;
+  unitType: string;
+  packSize: number;
+  salesPrice: number;
+  quantity: number;
+  id: number;
+}
