+++ src/app/models/profile-request.model.ts
@@ -1,0 +1,4 @@
+export interface ProfileRequest {
+  profile_did: number;
+  quantity: number;
+}
