+++ src/app/services/order-form.service.ts
@@ -1,65 +1,24 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Profile } from '../models/profile.model';
+import { OrderRequest } from '../models/order-request.model';
+import { OrderConfirmation } from '../models/order-confirmation.model';
+import { environment } from 'src/environments/environment';
+import { Order } from '../models/order.model';
 
 @Injectable({
   providedIn: 'root',
 })
 export class OrderFormService {
-  private apiUrl = 'https://uat-pffc.onrender.com/api/customers';
+  private apiUrl = environment.apiUrl;
 
   constructor(private http: HttpClient) {}
 
-  fetchCustomerData(customerId: string): Observable<any> {
-    return this.http.get<any>(`${this.apiUrl}/${customerId}/profiles`);
+  getOrderData(customerId: number): Observable<Order> {
+    return this.http.get<Order>(`${this.apiUrl}/customers/${customerId}/profiles`);
   }
 
-  placeOrder(
-    customerId: string,
-    orderData: any,
-  ): Observable<HttpResponse<any>> {
-    return this.http.post<any>(
-      `${this.apiUrl}/${customerId}/orders`,
-      orderData,
-      { observe: 'response' },
-    );
-  }
-
-  calculateTotal(products: Profile[]): number {
-    let total = 0;
-    products.forEach((product) => {
-      const quantity =
-        product.quantity !== undefined
-          ? parseFloat(product.quantity.toString())
-          : 0;
-      const price =
-        product.salesPrice !== undefined
-          ? parseFloat(product.salesPrice.toString())
-          : 0;
-      const packSize =
-        product.packSizePd !== undefined
-          ? parseFloat(product.packSizePd.toString())
-          : 1;
-      const lineTotal = quantity * packSize * price;
-      total += lineTotal;
-    });
-
-    const totalAmountSpan = document.getElementById(
-      'total-amount',
-    ) as HTMLSpanElement;
-    totalAmountSpan.textContent = total.toLocaleString('en-US', {
-      style: 'currency',
-      currency: 'USD',
-      minimumFractionDigits: 2,
-      maximumFractionDigits: 2,
-    });
-
-    const totalPriceInput = document.getElementById(
-      'total_price',
-    ) as HTMLInputElement;
-    totalPriceInput.value = total.toFixed(2);
-
-    return total;
+  placeOrder(customerId: number, orderData: OrderRequest): Observable<OrderConfirmation> {
+    return this.http.post<OrderConfirmation>(`${this.apiUrl}/customers/${customerId}/orders`, orderData);
   }
 }
