+++ src/environments/environment.development.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
   apiKey: 'devKey',
-  apiUrl: 'https://uat-pffc.onrender.com/api',
+  apiUrl: 'http://localhost:8080/api',
 };
